feat: 게시판서비스 로직에서 세션 검증 BoardService.java - 비즈니스 로직에서 세션에서 가져온 아이디로 검증하게 설정

+++ src/main/java/com/project/readers/service/BoardService.java
@@ -1,5 +1,6 @@
 package com.project.readers.service;
 
+import com.project.readers.config.SessionConfig;
 import com.project.readers.entity.BoardDTO;
 import com.project.readers.repository.BoardDAO;
 import lombok.RequiredArgsConstructor;
@@ -15,8 +16,12 @@ public class BoardService {
 
     private final BoardDAO boardDAO;
 
-    public void createBoard(BoardDTO boardDTO) {
+    public void createBoard(BoardDTO boardDTO) throws SecurityException {
+        if (SessionConfig.getSessionDTO().getId() == null) {
+            throw new SecurityException("회원만 작성할 수 있습니다.");
+        }
 
+        boardDTO.setCreator(SessionConfig.getSessionDTO().getId());
         boardDAO.createBoard(boardDTO);
     }
 
@@ -32,11 +37,24 @@ public class BoardService {
         return boardDAO.getBoardOne(boardNum);
     }
 
-    public void deleteBoard(Integer boardNum) {
-        boardDAO.deleteBoard(boardNum);
+    public void updateBoard(BoardDTO boardDTO) throws SecurityException {
+        String currentUserId = SessionConfig.getSessionDTO().getId();
+        String creatorId = boardDTO.getCreator();
+
+        if (currentUserId == null || !currentUserId.equals(creatorId)) {
+            throw new SecurityException("게시물의 작성자만 수정할 수 있습니다.");
+        }
+
+        // 수정할 내용을 게시물에 반영
+        boardDAO.updateBoard(boardDTO);
     }
 
-    public void updateBoard(BoardDTO boardDTO) {
-        boardDAO.updateBoard(boardDTO);
+    public void deleteBoard(Integer boardNum) throws SecurityException {
+        BoardDTO boardDTO = boardDAO.getBoardOne(boardNum);
+
+        if (!boardDTO.getCreator().equals(SessionConfig.getSessionDTO().getId())) {
+            throw new SecurityException("게시물의 작성자만 삭제할 수 있습니다.");
+        }
+        boardDAO.deleteBoard(boardNum);
     }
 }
